# api-doc.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: VinDr Lab API Document
  contact:
    name: VinDr Lab Development Team
  license:
    name: MIT
paths:
...
  /projects:
    get:
      description: get list by queried parameters
      operationId: fetchProjects
      parameters:
        - $ref: "#/components/parameters/authParam"
      responses:
        "200":
          description: get projects response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Project"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      description: get list by queried parameters
      operationId: createProject
      parameters:
        - $ref: "#/components/parameters/authParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Project"
      responses:
        "200":
          description: get projects response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Project"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /projects/{project_id}:
    get:
      description: get a Project by its id
      operationId: getProject
      parameters:
        - $ref: "#/components/parameters/authParam"
        - name: project_id
          in: path
          description: ID of Project
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: get project response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        $ref: "#/components/schemas/Project"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      description: update a Project by its id
      operationId: updateProject
      parameters:
        - $ref: "#/components/parameters/authParam"
        - name: project_id
          in: path
          description: ID of Project
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Project"
      responses:
        "200":
          description: get response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      description: delete a Project by its id
      operationId: deleteProject
      parameters:
        - $ref: "#/components/parameters/authParam"
        - name: project_id
          in: path
          description: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /projects/{project_id}/people:
    post:
      description: add people to project
      operationId: addPeopleToProject
      parameters:
        - $ref: "#/components/parameters/authParam"
        - name: project_id
          in: path
          description: ID of Project
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                people:
                  type: array
                  items:
                    $ref: "#/components/schemas/ProjectPerson"
      responses:
        "200":
          description: get project response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        properties:
                          people:
                            type: array
                            items:
                              $ref: "#/components/schemas/ProjectPerson"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      description: update people of project only 'people' field is accepted
      operationId: updatePeopleOfProject
      parameters:
        - $ref: "#/components/parameters/authParam"
        - name: project_id
          in: path
          description: ID of Project
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                people:
                  type: array
                  items:
                    $ref: "#/components/schemas/ProjectPerson"
      responses:
        "200":
          description: get project response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        properties:
                          people:
                            type: array
                            items:
                              $ref: "#/components/schemas/ProjectPerson"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /sessions:
    post:
      description: create session, client only have to pass data
      operationId: createSession
      parameters:
        - $ref: "#/components/parameters/authParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Session"
      responses:
        "200":
          description: get response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /sessions/{session_id}:
    get:
      description: get session
      operationId: getSession
      parameters:
        - $ref: "#/components/parameters/authParam"
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: get session
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/SessionItem"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
...
    authParam:
      name: Authorization
      in: header
      schema:
        type: string
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          enum:
            - value: 0
              name: Success
            - value: -1
              name: Failed
            - value: -2
              name: Invalid Data
            - value: -3
              name: Server Error
        message:
          type: string
        count:
          type: integer
          format: int32
          description: size of returned data
...
    Project:
      type: object
      required:
        - name
        - description
        - label_group_ids
        - reviewer_ids
        - project_owners_ids
        - workflow
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        created:
          type: integer
          format: int64
        creator_id:
          type: string
          format: uuid
        label_group_ids:
          type: array
          items:
            type: string
            format: uuid
        annotator_ids:
          type: array
          items:
            type: string
            format: uuid
        reviewer_ids:
          type: array
          items:
            type: string
            format: uuid
        project_owners_ids:
          type: array
          items:
            type: string
            format: uuid
        workflow:
          type: string
          enum: [SINGLE, TRIANGLE]
        document_link:
          type: string
        people:
          type: array
          items:
            $ref: "#/components/schemas/ProjectPerson"
        labeling_type:
          type: string
          enum: [3D, 2D]
...
    Session:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SessionItem"
        session_id:
          type: string
          format: uuid
        created:
          type: integer
          format: int64
    SessionItem:
      type: object
      properties:
        type:
          type: string
          enum: [TASK, STUDY]
        id:
          type: string
          format: uuid
        meta:
          type: object
          additionalProperties:
            type: string
...
    ProjectPerson:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        roles:
          type: array
          items:
            enum: [ANNOTATOR, REVIEWER, PROJECT_OWNER]
            type: string